requirements.txt
openai
instructor
python-pptx
pdfplumber
pydantic
streamlit
retrying
